Tiding up some files and test script
created some extra files to help with the structure
