+++ MP1/src/util.h
@@ -1,0 +1,7 @@
+#ifndef __UTIL_H_
+#define __UTIL_H_
+
+void concatenate(char a[], char b[]);
+void copy(char a[], char b[]);
+
+#endif // __UTIL_H_
